Refactor and update OloEngine core components
- Updated compiler detection macros for improved clarity and consistency.
- Enhanced debug break macros for better platform support.
- Refactored bit mask utility functions for better readability and maintainability.
- Improved atomic and simple flag implementations for better thread safety.
- Added profiling functionality to MemberList utility for performance tracking.
- Updated AudioThreadEvent and AudioThreadMessage tests to use new member variable naming conventions.
- Refactored SoundGraphAsset tests to utilize getter methods for better encapsulation.
- Fixed potential issues in StringUtils by adjusting delimiter handling.

+++ OloEngine/src/OloEngine/Audio/SoundGraph/Nodes/WavePlayer.h
@@ -461,7 +461,8 @@ namespace OloEngine::Audio::SoundGraph
 			{
 				for (u32 channel = 0; channel < m_AudioData.m_NumChannels; ++channel)
 				{
-					f32 sample = m_AudioData.GetSample(frame, channel);
+					// Safe cast: endFrame is already bounded by m_NumFrames (u32)
+					f32 sample = m_AudioData.GetSample(static_cast<u32>(frame), channel);
 					source.m_Channels.Push(sample);
 				}
 			}
